fix(pipeline): properly convert buffer into a string
Why is this change being made?
-------------------------------------
For a std::string of length N, the c-string equivalent has N+1 bytes
because it must end with a null terminator '\0'.

Our usage of NNG sent the first N bytes of a string of length N meaning
our data received by other nodes did not contain a null terminator
meaning they are not c-strings. This is not inherently wrong, but the
receiving nodes treated the received data as a c-string and constructed
a std::string from it. This leads to undefined behavior which likely was
the source of the prior bug.

What has changed?
-------------------------------------
Now we treat the data received from NNG as packed bytes of length N and
use that to initialize the std::string. This then prevents the prior
undefined behavior and the bug that came along with it.