feat(scrooge): Make scrooge use ipc for sent msgs
Why is this change being made?
---------------------------------
For scrooge to communicate messages from applications across the network
it needs to interface with outside applications. Now scrooge will open a
pipe in `/tmp/scrooge-input` and listen for ScroogeRequests protos.
Currently only SendMessageRequests are supported but soon this will also
be used for message verification.